+++ infrastructure/terraform/stacks/aws/outputs.tf
@@ -31,7 +31,6 @@ output "frontend_certificate_verification_dns_record" {
   description = "DNS record for Amplify SSL verification"
   value       = var.create_amplify_domain_association ? module.frontend_domain_association[0].certificate_verification_dns_record : null
 }
-<<<<<<< Updated upstream
 
 output "frontend_branches" {
   description = "Amplify frontend branches managed by Terraform"
@@ -43,5 +42,3 @@ output "frontend_branches" {
     }
   }
 }
-=======
->>>>>>> Stashed changes